Fixed wrong protobuf type reference in simulations
